+++ rust/connlib/tunnel/src/device_channel.rs
@@ -71,8 +71,8 @@ impl Device {
         dns_config: Vec<IpAddr>,
         callbacks: &impl Callbacks,
     ) -> Result<(), ConnlibError> {
+        // On Android / Linux we recreate the tunnel every time we re-configure it
         self.tun = Some(Tun::new(config, dns_config.clone(), callbacks)?);
-
         callbacks.on_set_interface_config(config.ipv4, config.ipv6, dns_config);
 
         if let Some(waker) = self.waker.take() {
@@ -113,6 +113,7 @@ impl Device {
         dns_config: Vec<IpAddr>,
         callbacks: &impl Callbacks,
     ) -> Result<(), ConnlibError> {
+        // On Windows we create the tunnel once and then we can reconfigure it in-place
         if self.tun.is_none() {
             self.tun = Some(Tun::new()?);
         }
